+++ libvast/src/system/pcap_reader_command.cpp
@@ -43,6 +43,7 @@ caf::message pcap_reader_command(const command& cmd, caf::actor_system& sys,
   using defaults_t = defaults::import::pcap;
   std::string category = defaults_t::category;
   reader_t reader{defaults::import::table_slice_type(sys, options),
+                  options,
                   get_or(options, category + ".read", defaults_t::read),
                   get_or(options, category + ".cutoff", defaults_t::cutoff),
                   get_or(options, category + ".max-flows",