+++ libvast/test/system/importer.cpp
@@ -103,7 +103,7 @@ struct importer_fixture : Base {
     namespace bf = format::zeek;
     auto stream = unbox(
       vast::detail::make_input_stream(artifacts::logs::zeek::small_conn));
-    bf::reader reader{vast::defaults::system::table_slice_type,
+    bf::reader reader{vast::defaults::system::table_slice_type, caf::settings{},
                       std::move(stream)};
     return this->self->spawn(system::source<bf::reader>, std::move(reader),
                              default_table_slice_builder::make, slice_size,