+++ libvast/vast/concept/parseable/core/rule.hpp
@@ -24,10 +24,47 @@ namespace vast {
 namespace detail {
 
 template <class Iterator, class Attribute>
+struct abstract_rule;
+
+template <class Iterator>
+struct abstract_rule<Iterator, unused_type> {
+  virtual ~abstract_rule() = default;
+  virtual abstract_rule* clone() const = 0;
+  virtual bool parse(Iterator& f, const Iterator& l, unused_type) const = 0;
+};
+
+template <class Iterator, class Attribute>
 struct abstract_rule {
   virtual ~abstract_rule() = default;
+  virtual abstract_rule* clone() const = 0;
   virtual bool parse(Iterator& f, const Iterator& l, unused_type) const = 0;
   virtual bool parse(Iterator& f, const Iterator& l, Attribute& a) const = 0;
+};
+
+template <class Parser, class Iterator, class Attribute>
+class rule_definition;
+
+template <class Parser, class Iterator>
+class rule_definition<Parser, Iterator, unused_type>
+  : public abstract_rule<Iterator, unused_type> {
+public:
+  explicit rule_definition(Parser p) : parser_(std::move(p)) {
+  }
+
+  rule_definition(const rule_definition& rhs) : parser_{rhs.parser_} {
+    // nop
+  }
+
+  rule_definition* clone() const override {
+    return new rule_definition(*this);
+  };
+
+  bool parse(Iterator& f, const Iterator& l, unused_type) const override {
+    return parser_(f, l, unused);
+  }
+
+private:
+  Parser parser_;
 };
 
 template <class Parser, class Iterator, class Attribute>
@@ -35,6 +72,14 @@ class rule_definition : public abstract_rule<Iterator, Attribute> {
 public:
   explicit rule_definition(Parser p) : parser_(std::move(p)) {
   }
+
+  rule_definition(const rule_definition& rhs) : parser_{rhs.parser_} {
+    // nop
+  }
+
+  rule_definition* clone() const override {
+    return new rule_definition(*this);
+  };
 
   bool parse(Iterator& f, const Iterator& l, unused_type) const override {
     return parser_(f, l, unused);
@@ -49,6 +94,57 @@ private:
 };
 
 } // namespace detail
+
+/// A type-erased parser which can store any other parser. This type exhibits
+/// value semantics and can therefore not be used to construct recursive
+/// parsers.
+template <class Iterator>
+class type_erased_parser : public parser<type_erased_parser<Iterator>> {
+public:
+  using abstract_rule_type = detail::abstract_rule<Iterator, unused_type>;
+  using rule_pointer = std::unique_ptr<abstract_rule_type>;
+  using attribute = unused_type;
+
+  template <class RHS>
+  static auto make_parser(RHS&& rhs) {
+    using rule_type = detail::rule_definition<RHS, Iterator, unused_type>;
+    return std::make_unique<rule_type>(std::forward<RHS>(rhs));
+  }
+
+  type_erased_parser() = default;
+
+  type_erased_parser(const type_erased_parser& rhs)
+    : parser_{rhs.parser_->clone()} {
+    // nop
+  }
+
+  template <class RHS, class = std::enable_if_t<!detail::is_same_or_derived_v<
+                         type_erased_parser, RHS>>>
+  type_erased_parser(RHS&& rhs)
+    : parser_{make_parser<RHS>(std::forward<RHS>(rhs))} {
+    static_assert(is_parser_v<std::decay_t<RHS>>);
+  }
+
+  type_erased_parser& operator=(const type_erased_parser& rhs) {
+    parser_.reset(rhs.parser_->clone());
+    return *this;
+  }
+
+  template <class RHS, class = std::enable_if_t<!detail::is_same_or_derived_v<
+                         type_erased_parser, RHS>>>
+  type_erased_parser& operator=(RHS&& rhs) {
+    static_assert(is_parser_v<std::decay_t<RHS>>);
+    parser_ = make_parser<RHS>(std::forward<RHS>(rhs));
+    return *this;
+  }
+
+  bool parse(Iterator& f, const Iterator& l, unused_type) const {
+    return parser_->parse(f, l, unused);
+  }
+
+private:
+  rule_pointer parser_;
+};
 
 /// A type-erased parser which can store any other parser.
 template <class Iterator, class Attribute = unused_type>
@@ -71,32 +167,24 @@ public:
   rule() : parser_{std::make_shared<rule_pointer>()} {
   }
 
-  template <
-    class RHS,
-    class = std::enable_if_t<
-      is_parser_v<std::decay_t<RHS>> && !detail::is_same_or_derived_v<rule, RHS>
-    >
-  >
-  rule(RHS&& rhs)
-    : rule{} {
+  template <class RHS, class = std::enable_if_t<
+                         is_parser_v<std::decay_t<
+                           RHS>> && !detail::is_same_or_derived_v<rule, RHS>>>
+  rule(RHS&& rhs) : rule{} {
     make_parser<RHS>(std::forward<RHS>(rhs));
   }
 
   template <class RHS>
   auto operator=(RHS&& rhs)
-    -> std::enable_if_t<is_parser_v<std::decay_t<RHS>>
-                        && !detail::is_same_or_derived_v<rule, RHS>> {
+    -> std::enable_if_t<is_parser_v<std::decay_t<
+                          RHS>> && !detail::is_same_or_derived_v<rule, RHS>> {
     make_parser<RHS>(std::forward<RHS>(rhs));
   }
 
-  bool parse(Iterator& f, const Iterator& l, unused_type) const {
+  template <class T>
+  bool parse(Iterator& f, const Iterator& l, T&& x) const {
     VAST_ASSERT(*parser_ != nullptr);
-    return (*parser_)->parse(f, l, unused);
-  }
-
-  bool parse(Iterator& f, const Iterator& l, Attribute& a) const {
-    VAST_ASSERT(*parser_ != nullptr);
-    return (*parser_)->parse(f, l, a);
+    return (*parser_)->parse(f, l, std::forward<T>(x));
   }
 
   const std::shared_ptr<rule_pointer>& parser() const {
@@ -164,4 +252,3 @@ auto ref(rule<Iterator, Attribute>& x) {
 }
 
 } // namespace vast
-