+++ libvast/vast/defaults.hpp
@@ -71,6 +71,22 @@ struct bgpdump {
   static constexpr auto read = shared::read;
 };
 
+/// Contains settings for the csv subcommand.
+struct csv {
+  /// Nested category in config files for this subcommand.
+  static constexpr const char* category = "import.csv";
+
+  /// Path for reading input events.
+  static constexpr auto read = shared::read;
+
+  static constexpr char separator = ',';
+
+  // TODO: agree on reasonable values
+  static constexpr std::string_view set_separator = ",";
+
+  static constexpr std::string_view kvp_separator = "=";
+};
+
 /// Contains settings for the json subcommand.
 struct json {
   /// Nested category in config files for this subcommand.
@@ -154,7 +170,7 @@ constexpr size_t max_events = 0;
 /// Contains settings for the zeek subcommand.
 struct zeek {
   /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "import.zeek";
+  static constexpr const char* category = "export.zeek";
 
   /// Path for writing query results.
   static constexpr auto write = shared::write;
@@ -163,16 +179,21 @@ struct zeek {
 /// Contains settings for the csv subcommand.
 struct csv {
   /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "import.csv";
+  static constexpr const char* category = "export.csv";
 
   /// Path for writing query results.
   static constexpr auto write = shared::write;
+
+  static constexpr char separator = ',';
+
+  // TODO: agree on reasonable values
+  static constexpr std::string_view set_separator = " | ";
 };
 
 /// Contains settings for the ascii subcommand.
 struct ascii {
   /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "import.ascii";
+  static constexpr const char* category = "export.ascii";
 
   /// Path for writing query results.
   static constexpr auto write = shared::write;
@@ -181,7 +202,7 @@ struct ascii {
 /// Contains settings for the json subcommand.
 struct json {
   /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "import.json";
+  static constexpr const char* category = "export.json";
 
   /// Path for writing query results.
   static constexpr auto write = shared::write;
@@ -190,7 +211,7 @@ struct json {
 /// Contains settings for the pcap subcommand.
 struct pcap {
   /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "import.pcap";
+  static constexpr const char* category = "export.pcap";
 
   /// Path for writing query results.
   static constexpr auto write = shared::write;